+++ lib/api/atproto.dart
@@ -176,6 +176,20 @@ abstract class Atproto {
         res.statusCode, json.decode(res.body));
   }
 
+  Future<Tuple2> getRecord(String repo, String collection, String rkey,
+      {String? cid}) async {
+    http.Response res = await api.get("com.atproto.repo.getRecord", params: {
+      "repo": repo,
+      "collection": collection,
+      "rkey": rkey,
+      "cid": cid,
+    }, headers: {
+      "Authorization": "Bearer ${api.session.accessJwt}"
+    });
+    return Tuple2<int, Map<String, dynamic>>(
+        res.statusCode, json.decode(res.body));
+  }
+
   Future<Tuple2> putRecord(
       String repo, String collection, String rkey, Map<String, dynamic> record,
       {bool? validate, String? swapRecord, String? swapCommit}) async {
@@ -283,5 +297,24 @@ abstract class Atproto {
     Map<String, dynamic> body =
         res.statusCode == 200 ? {} : json.decode(res.body);
     return Tuple2<int, Map<String, dynamic>>(res.statusCode, body);
+  }
+
+  Future<Tuple2> createReport(String reasonType, Map<String, dynamic> subject,
+      {String? reason}) async {
+    Map<String, dynamic> params = {
+      "reasonType": reasonType,
+      "subject": subject,
+    };
+    API.add(params, {
+      "reason": reason,
+    });
+    http.Response res = await api.post("com.atproto.moderation.createReport",
+        headers: {
+          "Content-Type": "application/json",
+          "Authorization": "Bearer ${api.session.accessJwt}"
+        },
+        body: json.encode(params));
+    return Tuple2<int, Map<String, dynamic>>(
+        res.statusCode, json.decode(res.body));
   }
 }