+++ test/flutter_bluesky_test.dart
@@ -1,14 +1,20 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bluesky/api/model/actor.dart';
+import 'package:flutter_bluesky/data/config.dart';
+import 'package:flutter_bluesky/util/post_util.dart';
 import 'package:http/http.dart' as http;
 import 'package:flutter_bluesky/api/model/feed.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_bluesky/flutter_bluesky.dart';
 import 'package:tuple/tuple.dart';
 import 'package:random_string/random_string.dart';
+import 'package:mockito/mockito.dart';
 
 // Integrated test code for FlutterBluesky.
 // Run local server with https://zenn.dev/tac519/articles/727fca3783010c
+
+// Mock Context
+// https://stackoverflow.com/questions/56277477/unit-testing-in-flutter-passing-buildcontext
 
 // XXX CHANGE consts with YOUR ENVIRONMENT.
 const testProvider = "http://localhost:2583";
@@ -22,6 +28,7 @@ const Map serverDescription = {
 };
 
 void main() {
+  config = Config({"timeout": 5000, "sleep": 500});
   FlutterBluesky plugin = FlutterBluesky(provider: testProvider);
 
   Future<Tuple2> login(String emailORhandle, String password) async {
@@ -30,7 +37,9 @@ void main() {
       plugin.api.session.setTokens(res.item2["did"], res.item2["handle"],
           res.item2["email"], res.item2["accessJwt"], res.item2["refreshJwt"]);
       await plugin.sessionAPI.profile();
+      setPlugin(plugin);
     }
+
     return res;
   }
 
@@ -128,8 +137,9 @@ void main() {
 
   test('post', () async {
     String text = randomAlphaNumeric(10);
+    await plugin.connect();
     await login(email, password);
-    await plugin.post(text);
+    await PostUtil.post(text, MockBuildContext(), files: []);
     Tuple2 res2 = await plugin.timeline();
     List feeds = res2.item2["feed"];
     bool exist = false;
@@ -156,7 +166,10 @@ void main() {
     List<Map>? images = [];
     Map element = {"image": res.item2["blob"], "alt": ""};
     images.add(element);
-    Tuple2 res2 = await plugin.post('post picture and text2', images: images);
+    Map<String, dynamic> record = {};
+    record["text"] = 'post picture and text2';
+    record["embed"] = {"\$type": "app.bsky.embed.images", "images": images};
+    Tuple2 res2 = await plugin.post(record);
     debugPrint("code: ${res2.item1}");
   });
 
@@ -176,3 +189,5 @@ void main() {
         avatar: res.item2["blob"]);
   });
 }
+
+class MockBuildContext extends Mock implements BuildContext {}